LC-43 Enable/Disable modules per category
