+++ conf/chat_modules.cfg
@@ -1,5 +1,15 @@
-## Приоритет по положению в файле.
+[gui_information]
+category = main
+
+[chats__gui]
+for = chats
+view = choose_multiple
+check_type = files
+check = modules/chats
+file_extension = False
+
 [chats]
 goodgame
 sc2tv
 twitch
+
