+++ conf/config.cfg
@@ -7,9 +7,6 @@ view = choose_single
 check_type = sections
 check = conf/translations.cfg
 
-[language]
-english
-
 [gui]
 show_hidden = True
 gui = True
@@ -17,3 +14,5 @@ on_top = True
 very_big_parameter_with_really_big_name_and_a_lot_of_not_needed_stuff = False
 reload
 
+[language]
+english
