+++ conf/messaging.cfg
@@ -1,8 +1,19 @@
+[gui_information]
+category = main
+
+[messaging__gui]
+for = messaging
+view = choose_multiple
+check_type = files
+check = modules/messaging
+file_extension = False
+
 [messaging]
+c2b
+logger
 blacklist
 mentions
-c2b
 df
 levels
 webchat
-logger
+
