+++ conf/translations.cfg
@@ -60,6 +60,12 @@ levels = Levels
 webchat = WebChat
 webchat.style.list_box = Available Styles
 
+messaging_modules = Message modules
+
+chat_modules = Chats
+
+logger = Logger
+
 [russian]
 menu.settings = Настройки
 menu.reload = Перезагрузить Веб Чат
