--- modules/chat_system.py
@@ -1,2 +1,0 @@
-SOURCE = 'sy'
-SOURCE_ICON = '/img/sources/lalka_cup.png'
