+++ modules/messaging/blacklist.py
@@ -9,7 +9,7 @@ class blacklist:
     users = {}
     words = {}
 
-    def __init__(self, conf_folder):
+    def __init__(self, conf_folder, **kwargs):
         # Dwarf professions.
         conf_file = os.path.join(conf_folder, "blacklist.cfg")
