--- modules/sc2tv.py
@@ -1,257 +1,0 @@
-import json
-import threading
-import time
-import re
-import requests
-import os
-import logging
-from ws4py.client.threadedclient import WebSocketClient
-from modules.helpers.parser import FlagConfigParser
-
-logging.getLogger('requests').setLevel(logging.ERROR)
-log = logging.getLogger('sc2tv')
-SOURCE = 'fs'
-SOURCE_ICON = 'http://funstream.tv/build/images/icon_home.png'
-
-
-class fsChat(WebSocketClient):
-    def __init__(self, ws, queue, nick, protocols=None, smiles=None):
-        super(self.__class__, self).__init__(ws, protocols=None)
-        # Received value setting.
-        self.source = SOURCE
-        self.queue = queue
-        self.nick = nick
-
-        self.channel_id = self.fsGetId()
-
-        self.smiles = smiles
-        self.smile_regex = ':(\w+|\d+):'
-
-        # Because funstream API is fun, we have to iterate the
-        #  requests in "proper" format:
-        #
-        #			42Iterator["command",{"params":"param"}]
-        # ex: 	420["/chat/join",{'channel':"stream/30000"}
-        # ex: 	421["/chat/join",{'channel':"stream/30000"}
-        # ex: 	429["/chat/join",{'channel':"stream/30000"}
-        # ex: 	4210["/chat/join",{'channel':"stream/30000"}
-        #
-        # Also, funstream API send duplicates of the messages
-        #  so we have to ignore the duplicates.
-        # We are doing so by creating special array which has
-        #  last N buffer of unique ID's
-        self.iter = 0
-        self.duplicates = []
-        self.users = []
-        self.bufferForDup = 20
-
-    def opened(self):
-        log.info("Websocket Connection Succesfull")
-
-    def closed(self, code, reason=None):
-        log.info("Websocket Connection Closed Down")
-
-    def allow_smile(self, smile, user_id):
-        allow = False
-
-        if smile['level'] == 0:
-            allow = True
-        # else:
-        #     user_match = False
-        #     for user_iter in self.users:
-        #         if user_iter['id'] == user_id:
-        #             user_match = True
-        #
-        #     if not user_match:
-        #         try:
-        #             user_id = {'id': user_id}
-        #             req_user = requests.post('http://funstream.tv/api/user/full', json=user_id)
-        #             if req_user.status_code == 200:
-        #                 req_user_answer = req_user.json()
-        #                 print "HelloWorld"
-        #                 print req_user_answer
-        #
-        #         except:
-        #             print "Unable to get smiles"
-        return allow
-
-    def received_message(self, mes):
-        # Funstream send all kind of different messages
-        # Some of them are strange asnwers like "40".
-        # For that we are trying to find real messages
-        #  which are more than 5 char length.
-        #
-        # Websocket has it's own type, so we serialise it to string.
-        message = str(mes)
-        if len(message) > 5:
-            # "Fun" messages consists of strange format:
-            # 	       43Iter{json}
-            # ex:      430{'somedata': 'somedata'}
-            # We need to just get the json, so we "regexp" it.
-            if re.findall('{.*}', message)[0]:
-                # If message does have JSON (some of them dont, dont know why)
-                #  we analyze the real "json" message.
-                message = json.loads(re.findall('{.*}', message)[0])
-                for dict_item in message:
-                    # SID type is "start" packet, after that we can join channels,
-                    #  at least I think so.
-                    if dict_item == 'sid':
-                        # "Funstream" has some interesting infrastructure, so
-                        #  we first need to find the channel ID from
-                        #  nickname of streamer we need to connect to.
-                        self.fsJoin()
-                        self.fsPing()
-                    elif dict_item == 'id':
-                        try:
-                            self.duplicates.index(message[dict_item])
-                        except:
-                            comp = {'source': self.source,
-                                    'source_icon': SOURCE_ICON,
-                                    'user': message['from']['name'],
-                                    'text': message['text'],
-                                    'emotes': []}
-                            if message['to'] is not None:
-                                comp['to'] = message['to']['name']
-                                if comp['to'] == self.nick:
-                                    comp['pm'] = True
-                            else:
-                                comp['to'] = None
-
-                            smiles_array = re.findall(self.smile_regex, comp['text'])
-                            for smile in smiles_array:
-                                for smile_find in self.smiles:
-                                    if smile_find['code'] == smile:
-                                        if self.allow_smile(smile_find, message['from']['id']):
-                                            comp['emotes'].append({'emote_id': smile, 'emote_url': smile_find['url']})
-
-                            self.queue.put(comp)
-                            self.duplicates.append(message[dict_item])
-                            if len(self.duplicates) > self.bufferForDup:
-                                self.duplicates.pop(0)
-
-    def fsGetId(self):
-        # We get ID from POST request to funstream API, and it hopefuly
-        #  answers us the correct ID of the channel we need to connect to
-        payload = "{'id': null, 'name': \"" + self.nick + "\"}"
-        request = requests.post("http://funstream.tv/api/user", data=payload)
-        if request.status_code == 200:
-            channel_id = json.loads(re.findall('{.*}', request.text)[0])['id']
-        else:
-            error_message = request.json()
-            if 'message' in error_message:
-                log.error("Unable to get channel ID. {0}".format(error_message['message']))
-            else:
-                log.error("Unable to get channel ID. No message available")
-            channel_id = None
-        return channel_id
-
-    def fsJoin(self):
-        # Because we need to iterate each message we iterate it!
-        iter = "42"+str(self.iter)
-        self.iter += 1
-
-        # Then we send the message acording to needed format and
-        #  hope it joins us
-        if self.channel_id:
-            join = str(iter) + "[\"/chat/join\", " + json.dumps({'channel': "stream/" + str(self.channel_id)}, sort_keys=False) + "]"
-            self.send(join)
-            log.info("Joined channel {0}".format(self.channel_id))
-
-    def fsPing(self):
-        # Because funstream is not your normal websocket they
-        #  have own "ping/pong" algorithm, and WE have to send ping.
-        #  Yes, I don't know why.
-        # We have to send ping message every 30 seconds, or funstream will
-        #  disconnect us. So we have to create separate thread for it.
-        # Dont understand why server is not sending his own pings, it
-        #  would be sooooo easier.
-        pingThr = pingThread(self)
-        pingThr.start()
-
-
-class pingThread(threading.Thread):
-    def __init__(self, ws):
-        threading.Thread.__init__(self)
-        self.daemon = "True"
-        # Using main websocket
-        self.ws = ws
-
-    def run(self):
-        # Basically, if we are alive we send every 30 seconds special
-        #  coded message, that is very hard to decode:
-        #
-        #      2
-        #
-        #  and they answer:
-        #
-        #      3
-        #
-        # No idea why.
-        while True:
-            self.ws.send("2")
-            time.sleep(30)
-
-
-class fsThread(threading.Thread):
-    def __init__(self, queue, socket, nick):
-        threading.Thread.__init__(self)
-        # Basic value setting.
-        # Daemon is needed so when main programm exits
-        # all threads will exit too.
-        self.daemon = "True"
-        self.queue = queue
-        self.socket = socket
-        self.nick = nick
-        self.smiles = []
-
-    def run(self):
-        # Let us get smiles for sc2tv
-        try:
-            smiles = requests.post('http://funstream.tv/api/smile')
-            if smiles.status_code == 200:
-                smiles_answer = smiles.json()
-                for smile in smiles_answer:
-                    self.smiles.append(smile)
-        except:
-            log.error("Unable to get smiles")
-
-        # Connecting to funstream websocket
-        ws = fsChat(self.socket, self.queue, self.nick, protocols=['websocket'], smiles=self.smiles,)
-        ws.connect()
-        ws.run_forever()
-
-
-class sc2tv:
-    def __init__(self, queue, python_folder):
-        log.info("Initializing funstream chat")
-
-        # Reading config from main directory.
-        conf_folder = os.path.join(python_folder, "conf")
-
-        conf_file = os.path.join(conf_folder, "sc2tv.cfg")
-        config = FlagConfigParser(allow_no_value=True)
-        if not os.path.exists(conf_file):
-            config.add_section('config')
-            config.set('config', 'socket/hidden', 'ws://funstream.tv/socket.io/')
-            config.set('config', 'nick', 'adolfra')
-
-            config.write(open(conf_file, 'w'))
-
-        self.conf_params = {'folder': conf_folder, 'file': conf_file,
-                            'filename': ''.join(os.path.basename(conf_file).split('.')[:-1]),
-                            'parser': config}
-
-        config.read(conf_file)
-        # Checking config file for needed variables
-        config_tag = 'config'
-        socket = config.get_or_default(config_tag, 'socket', 'ws://funstream.tv/socket.io/')
-        nick = config.get_or_default(config_tag, 'nick', 'adolfra')
-
-        # If any of the value are non-existent then exit the programm with error.
-        if (socket is None) or (nick is None):
-            log.critical("Config for funstream is not correct!")
-            exit()
-
-        # Creating new thread with queue in place for messaging tranfers
-        fs = fsThread(queue, socket, nick)
-        fs.start()
